Propagate tags to intermediate targets
The top-level "cue_consolidated_..." and "cue_exported_..." macros
wrap rules that accept a "tags" attribute, and both sets of rules
generate a few intermediate targets to collect runfiles and feed them
to a launcher program that eventually invokes the "cue" tool. If we
don't propagate tags such as "manual" that are present on the
top-level targets to these intermediate targets, then the intermediate
targets may be included unintentionally in wildcard target patterns,
causing failures when those targets' dependencies require a particular
configuration to be active.

In order to ensure that these intermediate targets behave equivalently
to the targets that motivated them, propagate the same tags specified
for the latter to the former.