Ensure that annotated tag content is available
When preparing the notes to accompany a release that we issue upon
receiving a pushed tag, we try to include both the subject and body of
an annotation attached to the motivating Git tag. The
"actions/checkout" GitHub Action does not fetch tag annotations by
default, and coercing it to do so (in its current form) would require
fetching all of the Git repository's history.

Work around this problem by continuing to fetch the repository history
shallowly, but then fetching the repository's tags separately
afterward. Doing so makes these annotations available.